+++ app/Main.hs
@@ -1,0 +1,6 @@
+module Main (main) where
+
+import Codec.Compression.ShannonFano
+
+main :: IO ()
+main = undefined